+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add option `discord_custom_emojis` to enable our own Discord emojis for factions and professions. Enabled by default.
+- Added `jack` / `legchopper` command to track Jack Legchopper and his clones
 
 ### Changed